consolidating codex.ts and useCodex.ts
